third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/15.3.5.2-expected.txt

PASS 'use strict'; var f = fn(); delete f.prototype threw exception of type TypeError.
PASS var f = fn(); delete f.prototype is false
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
